Refactor: Tokens now carry their text
This means that after parsing the file in ParseEntireFile, subsequent operations (such as FindMatchedPairs) don't need to access to the array of file lines. This in turn means that we can factor out the parsing functionality and reuse it for other kinds of operations -- because now the parse can simply return a list of tokens.