Refactor: Prepare for finding single matching brace.
Up until now, we've been going about brace matching in the completely wrong way. We're supposed to find the one brace that matches the one at the location of the caret, we're supposed to inform the AuthoringSink of that one pair, and we're supposed to set the AuthoringSink's FoundMatchingBrace property.

We've been doing none of those things. Instead, we've been informing the AuthoringSink of every pair in the file, and we've not been setting FoundMatchingBrace.

In this commit, we separate the process of finding all the pairs (existing method, now slightly modified and renamed to FindBraceMatches) from the process of deciding which is the right pair (new method FindMatchForBrace, which is not yet implemented and returns null). We keep the unit-test-facilitating event (now renamed to BraceMatchFound), so we can continue to verify that we find all the right pairs in a file.

In the next commit, we'll implement FindMatchForBrace, after which brace matching should actually work.

Kudos to James Gregory and his blog post http://www.jagregory.com/writings/brace-matching-and-your-language-service/ for the right approach.