Experiment with tracking line numbers
The base language service tracks line numbers. A derived language service works completely in terms of character indices, and if it every needs line and column numbers, it can ask the Source object to convert.

This makes it hard to unit test functionality (such as brace highlighting) that depends on line/column numbers, because the language service gets the Source object from the view:

    Source source = GetSource(req.View);

... and we don't have a view.

In this commit on the "line-numbers" experimental branch, we experimeted with tracking line numbers ourself. It's not that hard, but it does require us to treat CRLF and a single line break, and to know how the view treats unmatched CRs and LFs.

I'm going to put this aside for now and move on to some other features.