Enable background parsing
There appears to be a bug in the language service infrastructure, the effect of which is that you never get a call to LanguageService.ParseSource with ParseReason.Check. Apparently the LastParseTime property of the Source object is set to 0x7fffffff and never reset. So we reset it in the OnIdle method.