Partial fix for #10: Parse error reported when assigning value to $NodeDeallocationOption
We fix this by allowing a keyword on the right of an assignment, and adding a rule that this is only allowed when assigning to $NodeDeallocationOption.

In the next commit, we'll complete the fix by implementing a rule that says that _only_ these keywords are allowed when assigning to $NodeDeallocationOption.