Fix #12: DRY out rule test classes
The test methods in those classes all have the same structure.

We can't put the entire test method into a base class, because the test method is decorated with [MemberData("TestCases")], and Xunit insists that a member of that name exist on the base class. The best we can do is factor the bodies of the test methods into a common helper in the base class.

This really is worth doing. In the course of doing it we uncovered a bug in one of the UnknownMethodName test cases. The bug was missed because the test method in that class didn't have quite the same structure as the methods in the other rule test classes. Fortunately it was a test bug, not a product bug.