Fix #23: Reset tokenInfo.Trigger before OR-ing in new triggers
The language service seems to reuse the same `TokenInfo` object each time it calls the `Scanner`. So be sure to set the `Trigger` property to `TokenTriggers.None` before OR-ing in the appropriate triggers, if any, for the current token.

Failing to do this meant that the triggers were accumulated from call to call. So, for example, once a `ParenClose` token was seen, all subsequent tokens had the `TokenTriggers.MatchBraces` flag set. This caused the language service to do more work than it needed to (attempting to match braces after every token).